requirements.txt
APScheduler==3.3.1
Automat==0.6.0
Django==1.11.3
Flask-Login==0.4.0
Flask==0.12.1
Jinja2==2.9.6
MarkupSafe==1.0
Pillow==4.1.1
PyDispatcher==2.0.5
PyExecJS==1.4.0
Pygments==2.2.0
Scrapy==1.4.0
incremental==17.5.0
Twisted==17.1.0
Werkzeug==0.12.1
WsgiDAV==2.2.1
appdirs==1.4.3
appnope==0.1.0
asn1crypto==0.22.0
async-timeout==1.2.1
attrs==17.2.0
beautifulsoup4==4.6.0
bitarray==0.8.1
cchardet==2.1.0
cffi==1.10.0
chardet==3.0.2
click==6.7
constantly==15.1.0
crcmod==1.7
cryptography==1.9
cssselect==1.0.1
dateutils==0.6.6
decorator==4.0.11
defusedxml==0.5.0
django-cors-headers==2.1.0
django-extensions==1.8.0
django-filter==1.0.4
django-rest-swagger==2.1.2
djangorestframework==3.6.3
elasticsearch==5.4.0
fake-useragent==0.1.7
html2text==2016.9.19
html5lib==0.999999999
idna==2.5
ipython-genutils==0.2.0
ipython==5.4.1
itsdangerous==0.24
jedi==0.10.2
lxml==3.8.0
mongoengine==0.13.0
multidict==2.1.5
mysqlclient==1.3.10
numpy==1.13.1
olefile==0.44
oss2==2.3.3
packaging==16.8
pandas==0.20.3
parsel==1.2.0
pexpect==4.2.1
pickleshare==0.7.4
pika==0.10.0
prompt-toolkit==1.0.14
ptyprocess==0.5.1
pyOpenSSL==17.0.0
pyaml==16.12.2
pyasn1-modules==0.0.8
pyasn1==0.2.3
pycares==2.1.1
pycparser==2.17
pycrypto==2.6.1
pymongo==3.4.0
pymysql==0.7.11
pyparsing==2.2.0
pyquery==1.2.17
python-dateutil==2.6.0
pytz==2017.2
queuelib==1.4.2
redis==2.10.5
requests==2.18.1
scrapyd==1.2.0
service-identity==17.0.0
simplegeneric==0.8.1
simplejson==3.11.1
six==1.10.0
sqlalchemy==1.1.12
sqlbuilder==0.7.10.11
tblib==1.3.2
timeout-decorator==0.4.0
tornado==4.5.1
traitlets==4.3.2
tzlocal==1.4
u-msgpack-python==2.4.1
uWSGI==2.0.15
urllib3==1.21.1
virtualenv==15.1.0
virtualfish==1.0.5
w3lib==1.17.0
wcwidth==0.1.7
webencodings==0.5.1
websocket-client==0.40.0
wechatpy==1.5.1
whenpy==0.3.0
wmi==1.4.9
PyExecJS==1.4.0
Fabric3==1.13.1.post1
pymongo==3.4.0
certifi==2015.04.28
